requirements.txt
discord
ollama
dotenv
websocket-client
Pillow
requests
beautifulsoup4
lxml